extracted bot credentials
