+++ src/main/java/com/gainetdin/telegram/DebtAccounterBotApplication.java
@@ -2,8 +2,10 @@ package com.gainetdin.telegram;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class DebtAccounterBotApplication {
 
 	public static void main(String[] args) {
