+++ src/main/java/com/gainetdin/telegram/bot/BotCredentials.java
@@ -1,0 +1,26 @@
+package com.gainetdin.telegram.bot;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "bot")
+public class BotCredentials {
+
+    private String name;
+    private String token;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+}
