+++ src/main/java/com/gainetdin/telegram/facade/MessageUpdatingFacade.java
@@ -1,0 +1,4 @@
+package com.gainetdin.telegram.facade;
+
+public class MessageUpdatingFacade {
+}
